session

+++ platformWeb/src/main/java/com/gennlife/platform/util/RedisUtil.java
@@ -164,7 +164,7 @@ public class RedisUtil {
         deleteUser(uid);
         try {
             AllDao.getInstance().getSessionDao().deleteByUid(uid);
-            AllDao.getInstance().getSessionDao().deleteBySessionID(sessionID);
+            if(sessionID!=null)AllDao.getInstance().getSessionDao().deleteBySessionID(sessionID);
     }
         catch (Exception e)
         {
